Suppress needs to have dll-interface warnings for std::weak_ptr<T>
The warning is nonpractical; Even if the warning is solved, which can be
be achieved through using a different shared pointer implementation, the fact
remains that memory layouts may differ between the library and the consumer.
Shared pointers are an integral part of the C++11 Language Mapping, users are
required to ensure binary compatibility.

Signed-off-by: Jeroen Koekkoek <jeroen@koekkoek.nl>