Fix for reading more entries than allowed in a bounded sequence
While writing a bounded sequence with more entries than allowed
would throw an exception, reading more entries would not cause
an error.
While the number of entries will still need to be read, to allow
for the position index to be updated correctly for complex types
(strings, sequences, etc.) the end result will be truncated to
the bound if the bound is exceeded.

Signed-off-by: Martijn Reicher <martijn.reicher@adlinktech.com>