Always set the discrimant value to the one read
This forces the discriminant value to be the actual one read from the
CDR, not the default value associated with the setter we used.  E.g.:

  union U switch (long) {
    case 1: case 2: long a;
    default: string b;
  };

setting a value with a(...) will initialize the discriminant to 1, even
when the intended discrimant value is 2; the same problem exists with
the default case, as well as when where there is no "default:" but the
discrimant value doesn't map to any case.

Note that this fix is suboptimal: for the vast majority of the unions,
this doesn't come into play and this merely sets the discrimant value
the same value twice.

Signed-off-by: Erik Boasson <eb@ilities.com>