C++ serdata hash: mix in serdata_basehash
Each type gets its own specialised functions, and consequently the tkmap
hash table (in the core) will never consider two samples of different
types as equal.

The serdata hash is currently computed as the MD5 hash of the key value,
and so without mixing in "serdata_basehash" all samples with the same
key value end up with the same hash despite not comparing as equal.  In
other words, if there are many topics there are many hash collisions.
The "hopscotch" hash table used by tkmap blows up if it is fed too many
collisions.

Mixing in "serdata_basehash" addresses that problem.

Signed-off-by: Erik Boasson <eb@ilities.com>