Bump minimum CMake version to 3.16
Support for EXPORT_PROPERTIES was added in CMake 3.12, so either the
minimum version required needs to be bumped, or this feature needs to be
emulated.  Ubuntu 22.04 ships with CMake 3.16, is over two years old now
and also the minimum required for ROS 2 Foxy, so that seems like a
reasonable new minimum version.

Signed-off-by: Erik Boasson <eb@ilities.com>