Fix for issue #293
This issue was caused by the register_types function not looking
at the declarator in a typedef to see whether the arrays include
is required. When the typedef was used, then the include was
put in correctly.
The fix was to look at the declarators of typedefs by themselves
too.

Signed-off-by: Martijn Reicher <martijn.reicher@zettascale.tech>