Free type map/info only if successfully generated
On failure generating the type info and type map, the code using these
blobs when writing out the traits would still free the typemap and
typeinfo blobs, even though the contents of the memory would be
undefined.  This changes the error handling so this is done only iff the
type info and type map were generated correctly.

Signed-off-by: Erik Boasson <eb@ilities.com>