Multi-hop latency measurement example
This adds a simple tool for doing latency measurements across multiple hops.  It assumes
the clocks are synchronised to a high degree so that one-way latencies can be computed
directly.

It can operate with a number of different types, all very simple:

    struct Hop8 {
      uint32 seq;
      octet z[8 - 4];
    };

and variants where the total size is 128, 1k, 8k and 128k bytes.  Each process takes a
stage, with the source publishing in partition Pj, the sink subscribing in partition Pj
and the forwarders subscribing in Pj and publishing in Pk, where j is the stage argument
and k = j+1.

Each process additionally subscribes to "junk data" and optionally publishes samples at
randomised intervals with a configurable average rate.

Signed-off-by: Erik Boasson <eb@ilities.com>