Multi-topic latency measurement example
This adds a simple tool for doing latency measurements between processes when writing
multiple topics.  It assumes the clocks are synchronised to a high degree so that one-way
latencies can be computed directly.

It can operate with a number of different types, all very simple:

    struct Hop128 {
      @key uint32 key;
      uint32 seq;
      octet z[128 - 8];
    };

and variants where the total size is 8, 1k, 8k and 128k bytes.

A process is either a source of data, writing an each instance for each of N topics every
10ms, or it is a sink, recording for each received sample the latency and the topic.  It
can publish these N samples as quickly in succession as possible, or it can do it while
requesting 100us sleeps in between.

Signed-off-by: Erik Boasson <eb@ilities.com>