Switch union std::variant access to being by index
Accessing std::variant by index means the type list for the variant
doesn't have to be de-duped, saving a chunk of logic and ensuring
bounded and unbounded strings can be mixed in unions with no problems.

Preserve access by type for use with boost::variant, where it is the
only option.

Signed-off-by: Jim Hague <jim.hague@acm.org>