Extend type deduplication in variants to check cpp11 type
Currently code generated for unions with a mixture of bounded and
unbounded strings fails to compile. By default, the union generates a
std::variant or boost::variant, accessing by type. The union member
types are de-duplicated to ensure all types in the variant are unique.

The deduplication regards unbounded strings and bounded strings of
different bounds length as distinct types, so does not de-duplicate.
However, the default templates give std::string for both bounded and
unbounded strings of all lengths, and a variant is generated with
multiple members of the same type, and access via type fails.

Counter this by adding a final de-duplication step which does additional
de-duplication on cpp11 type.

Signed-off-by: Jim Hague <jim.hague@acm.org>