Use XCDR1 instead of basic CDR in C++ ser/deser
The XCDR1 support in the core library changed the behaviour of the code that picked the
default data representation: it now looks at the "data type properties" in the sertype.
The C++ code does not correctly set this, and the consequence is that the default encoding
for an appendable C++ topic changed from XCDR2 to XCDR1.

The C++ however distinguishes between "basic CDR" (the old CORBA standard) and XCDR1 in
its serializer. (The two have the same serialization for types supported under the old
standard, and so this distinction is not meaningful.) Instead of mapping the encoding to
XCDR1, it mapped it to "basic CDR", and in "basic CDR" mode, it refuses to serialize
anything added by XTypes, such as an appendable struct. This causes an attempt to write
such data to fail.

The default encoding is for, e.g., an appendable struct is still incorrect.

Signed-off-by: Erik Boasson <eb@ilities.com>