Do not pad XCDR1 non-mutable optional to 4 bytes
Padding out to 4 bytes for alignment of the following parameter header is wrong, but in
some sense understandable. Padding out to 4 bytes if the following object in the stream
does not require 4 byte alignment makes no sense whatsoever.

The change for padding out the headers in mutable lists accidentally changed the padding
for optionals in non-mutable types. This fixes that.

Signed-off-by: Erik Boasson <eb@ilities.com>