Fix handling of optional fields in mutable structs
The alignment handling for optional fields in mutable types depends on whether the field
is present but uses the presence of the field in the application representation before it
has been set by deserializing the value. This causes it to pop a alignment offset off an
internal stack without having pushed it onto the stack.

Signed-off-by: Erik Boasson <eb@ilities.com>